+++ src/lib/declarationHandler.js
@@ -61,7 +61,7 @@ const CLASSES_OF_TRUTH = [
     'DagGlobals',
     'System',
     'TrafficControl',
-    'SSH'
+    'SSHD'
 ];
 
 /**