+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changes to this project are documented in this file. More detail and links can b
 - Add support for HTTPD class [Issue 50](https://github.com/F5Networks/f5-declarative-onboarding/issues/50)
 - Add support for consoleInactivityTimeout in System class
 - Add support for hostname in System class
+- AUTOTOOL-747: Add Declarative Onboarding analytics reporting to F5
 ### Fixed
 - Fix bug in which DO was unable to use management network for SnmpTrapDestination
 - Fix bug in which DO creates incomplete RADIUS authentication configuration